+++ src/App.tsx
@@ -5,6 +5,8 @@ import {
   onMount,
   onCleanup,
   Show,
+  batch,
+  createEffect,
 } from "solid-js";
 
 import Table, { Column } from "./Table";
@@ -53,12 +55,29 @@ const App: Component = () => {
 
   const { isOpen, onOpen, onClose } = createOpen();
 
+  const [scroll, setScroll] = createSignal<
+    ((index: number) => void) | undefined
+  >(undefined);
+
   const events = createMemo(() =>
     searchEvents({
       timestampRange: timestampRange(),
       fulltext: fulltext(),
     })
   );
+
+  const scrollTo = (index: number) => {
+    const scrollFunc = scroll();
+    if (scrollFunc != null) {
+      scrollFunc(index);
+    }
+  };
+
+  createEffect(() => {
+    events();
+    // scroll to index 0 when events are filtered
+    scrollTo(0);
+  });
 
   const handleFake = (amount: number) => {
     setEvents(randomEvents(new Date(Date.now()), amount));
@@ -83,11 +102,12 @@ const App: Component = () => {
     setEvents(events);
   };
 
-  const handleContext = (event: DeltaChatEvent) => {
-    setTimestampRange({
-      start: event.ts - CONTEXT_TIME,
-      end: event.ts + CONTEXT_TIME,
+  const handleContext = (selected: DeltaChatEvent) => {
+    batch(() => {
+      setFulltext("");
+      setTimestampRange({ start: null, end: null });
     });
+    scrollTo(selected.id);
   };
 
   onMount(() => {
@@ -102,13 +122,16 @@ const App: Component = () => {
     <AppContainer>
       <Header>
         <div class="flex flex-col gap-2 bg-slate-200 px-2 py-2">
-          <div class="flex flex-row gap-2 bg-slate-200">
-            <TextInput
-              placeholder="Search"
-              value={fulltext}
-              setValue={setFulltext}
-              debounce={500}
-            />
+          <div class="flex flex-row justify-between gap-2 bg-slate-200">
+            <div class="flex w-1/2 flex-row justify-between">
+              <TextInput
+                placeholder="Search"
+                value={fulltext}
+                setValue={setFulltext}
+                debounce={500}
+              />
+              <div class="text-lg">{events().length}</div>
+            </div>
             <Show
               when={!isOpen()}
               fallback={<Button onClick={onClose}>Less</Button>}
@@ -138,6 +161,9 @@ const App: Component = () => {
           infoModal={(props) => (
             <EventInfo {...props} onContext={handleContext} />
           )}
+          setScroll={(scroll) => {
+            setScroll(() => scroll);
+          }}
         />
       </Content>
     </AppContainer>