+++ src/Table.tsx
@@ -1,6 +1,14 @@
-import { createSignal, Component, For, Accessor, createMemo } from "solid-js";
+import {
+  createSignal,
+  Component,
+  For,
+  Accessor,
+  createMemo,
+  createEffect,
+  onMount,
+} from "solid-js";
 
-import { createVirtualizer } from "./solid-virtual";
+import { createVirtualizer, VirtualItem } from "./solid-virtual";
 import { createOpen } from "./createOpen";
 
 export type Column<T> = {
@@ -12,23 +20,42 @@ export type Column<T> = {
 function Row<T>(props: {
   columns: Column<T>[];
   data: Accessor<T[]>;
-  index: number;
-  onSelect: (value: Accessor<T | undefined>) => void;
+  virtualItem: VirtualItem<unknown>;
+  onSelect: (value: T) => void;
+  onInfo: (value: T) => void;
+  isSelected: (value: T) => boolean;
 }) {
   const value = createMemo(() => {
-    return props.data()[props.index];
+    return props.data()[props.virtualItem.index];
   });
 
   return (
     <div
-      class="flex w-full flex-row justify-start gap-2.5"
-      onClick={() => {
-        props.onSelect(value);
+      class="absolute top-0 left-0 flex w-full items-center justify-center"
+      classList={{
+        "bg-slate-200": props.virtualItem.index % 2 === 0,
+        "bg-slate-400": props.isSelected(value()),
       }}
+      style={{
+        height: `${props.virtualItem.size}px`,
+        transform: `translateY(${props.virtualItem.start}px)`,
+      }}
+      onClick={() => {
+        props.onSelect(value());
+      }}
+      onDblClick={() => {
+        props.onInfo(value());
+      }}
+      // onContextMenu={() => {
+      //   // this should be a long-press on mobile
+      //   props.onInfo(value());
+      // }}
     >
-      <For each={props.columns}>
-        {(column) => <Cell value={value} column={column} />}
-      </For>
+      <div class="flex w-full flex-row justify-start gap-2.5">
+        <For each={props.columns}>
+          {(column) => <Cell value={value} column={column} />}
+        </For>
+      </div>
     </div>
   );
 }
@@ -50,26 +77,26 @@ function Cell<T>(props: { column: Column<T>; value: Accessor<T> }) {
   );
 }
 
-type InfoModal<T> = Component<{
-  value: Accessor<T | undefined>;
+export type InfoModalProps<T> = {
+  selected: Accessor<T | undefined>;
   onClose: () => void;
   isOpen: Accessor<boolean>;
-}>;
+};
+
+type InfoModal<T> = Component<InfoModalProps<T>>;
 
 function Table<T>(props: {
   columns: Column<T>[];
   data: Accessor<T[]>;
   infoModal: InfoModal<T>;
+  setScroll: (scroll: (index: number) => void) => void;
 }) {
   const { isOpen, onClose, onOpen } = createOpen();
-  const [selectedValue, setSelectedValue] = createSignal<T | undefined>(
-    undefined
-  );
 
   const handleClose = () => {
-    setSelectedValue(undefined);
     onClose();
   };
+
   // it's important for this ref not to be height 100% as
   // this breaks virtualizer
   let scrollParentRef: HTMLDivElement | undefined;
@@ -83,9 +110,39 @@ function Table<T>(props: {
     })
   );
 
+  createEffect(() => {
+    const virtualizer = rowVirtualizer();
+    props.setScroll((index) => {
+      virtualizer.scrollToIndex(index, { align: "center" });
+    });
+  });
+
+  const [selected, setSelected] = createSignal<T | undefined>(undefined);
+
+  const handleSelect = (selected: T) => {
+    setSelected(() => selected);
+  };
+
+  const handleInfo = (selected: T) => {
+    setSelected(() => selected);
+    onOpen();
+  };
+
+  const isSelected = (record: T) => {
+    const current = selected();
+    if (current == null) {
+      return false;
+    }
+    return current === record;
+  };
+
   return (
-    <div class="flex h-full flex-col">
-      {props.infoModal({ value: selectedValue, onClose: handleClose, isOpen })}
+    <div class="flex h-full select-none flex-col">
+      {props.infoModal({
+        selected: selected,
+        onClose: handleClose,
+        isOpen,
+      })}
       <div class="flex w-full flex-row justify-start gap-2.5">
         <For each={props.columns}>
           {(column) => (
@@ -107,30 +164,16 @@ function Table<T>(props: {
           }}
         >
           <For each={rowVirtualizer().getVirtualItems()}>
-            {(virtualItem) => {
-              return (
-                <div
-                  class="absolute top-0 left-0 flex w-full items-center justify-center"
-                  classList={{
-                    "bg-slate-200": virtualItem.index % 2 === 0,
-                  }}
-                  style={{
-                    height: `${virtualItem.size}px`,
-                    transform: `translateY(${virtualItem.start}px)`,
-                  }}
-                >
-                  <Row
-                    data={props.data}
-                    columns={props.columns}
-                    index={virtualItem.index}
-                    onSelect={(value: Accessor<T | undefined>) => {
-                      onOpen();
-                      setSelectedValue(value);
-                    }}
-                  />
-                </div>
-              );
-            }}
+            {(virtualItem) => (
+              <Row
+                data={props.data}
+                columns={props.columns}
+                virtualItem={virtualItem}
+                onInfo={handleInfo}
+                onSelect={handleSelect}
+                isSelected={isSelected}
+              />
+            )}
           </For>
         </div>
       </div>